requirements.txt
aerich==0.9.2
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncclick==8.3.0.7
asyncssh==2.21.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.3
dictdiffer==0.9.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iso8601==2.1.0
loguru==0.7.3
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
sniffio==1.3.1
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
win32_setctime==1.2.0
yarl==1.22.0
